layout: Split `SameFormattingContextBlock` into its own struct
This allows splitting up `flow/mod.rs` which is getting huge and also
prepare the way for starting to cache hoisted absolutes on
`SameFormattingContextBlock`s that establish their containing block.
